src/ai_src/server/requirements.txt
flask
pydub
numpy
google-cloud-speech
